issue-34 refactor: moved HashingErrors to their dedicated packages
